requirements.txt
openai
flask
flask-cors
python-dotenv
google-generativeai
